Tasks completed for part4. Header files and implementations created for each object type.
