+++ .replit
@@ -1,2 +1,2 @@
 language = "cpp"
-run = "clang++ -std=c++17 -Weverything -Wno-padded -Wno-c++98-compat -Wno-missing-prototypes -Wno-deprecated -Wno-exit-time-destructors main.cpp && ./a.out"
+run = "clang++ -std=c++17 -Weverything -Wno-padded -Wno-c++98-compat -Wno-missing-prototypes -Wno-deprecated -Wno-exit-time-destructors main.cpp RugbyTeam.cpp SoccerTeam.cpp Staff.cpp TeamProgram.cpp TrainingComplex.cpp Wrappers.cpp && ./a.out"
